refactor: divide TransactionDetails component
